added prevention for sql injection and escaped all data going to database

+++ CMS/admin/partials/view_all_users.php
@@ -65,11 +65,17 @@
 	<?php
 
 		if(isset($_GET['delete'])) {
-			$delete_user_id = $_GET['delete'];
-			$query = "DELETE FROM users WHERE user_id = {$delete_user_id} ";
-			$delete_query = mysqli_query($connection, $query);
-			confirm($delete_query);
-			header("Location: users.php");
+
+			if(isset($_SESSION['user_role'])){
+
+				if($_SESSION['user_role'] == 'admin'){
+					$delete_user_id = mysqli_real_escape_string($_GET['delete']);
+					$query = "DELETE FROM users WHERE user_id = {$delete_user_id} ";
+					$delete_query = mysqli_query($connection, $query);
+					confirm($delete_query);
+					header("Location: users.php");
+				}
+			}
 		}
 //							Set approved and unapproved in comments table
